requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
decorator
entrypoints
fonttools
gitdb
GitPython
idna
importlib-metadata
Jinja2
jsonschema
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
mdurl
numpy
packaging
pandas
patsy
Pillow
plotly
plotly-express
protobuf
pyarrow
pydeck
Pygments
Pympler
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
requests
rich
scipy
semver
six
smmap
statsmodels
streamlit
tenacity
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
zipp
